+++ syngular/field.py
@@ -36,12 +36,14 @@ class Field(object):
 
     def __call__(self, other):
         """Cast to field."""
-        if self.name == "mpc":
+        if self.name in ["mpc", "C"]:
             return mpmath.mpc(mpmath.mpmathify(other))
-        elif self.name == "padic":
+        elif self.name in ["padic", "Qp"]:
             return PAdic(other, self.characteristic, self.digits)
-        elif self.name == "finite field":
+        elif self.name in ["finite field", "Fp"]:
             return ModP(other, self.characteristic)
+        elif self.name in ["rational", "Q"]:
+            return Q(other)
         else:
             raise NotImplementedError
 
@@ -51,7 +53,7 @@ class Field(object):
 
     @name.setter
     def name(self, value):
-        if value not in ['mpc', 'gaussian rational', 'finite field', 'padic']:
+        if value not in ['rational', 'Q', 'mpc', 'C', 'gaussian rational', 'finite field', 'Fp', 'padic', 'Qp']:
             raise Exception("Field must be one of 'mpc', 'gaussian rational', 'finite field', 'padic'.")
         else:
             self._name = value
@@ -69,7 +71,7 @@ class Field(object):
 
     @property
     def digits(self):
-        if self.name == 'mpc':
+        if self.name in ['mpc', 'C']:
             return mpmath.mp.dps
         else:
             return self._digits
@@ -78,7 +80,7 @@ class Field(object):
     def digits(self, value):
         if value < 0:
             raise Exception("Digits must be positive.")
-        elif self.name == 'mpc':
+        elif self.name in ['mpc', 'C']:
             mpmath.mp.dps = value
         else:
             self._digits = value
@@ -88,7 +90,7 @@ class Field(object):
 
     @property
     def is_algebraically_closed(self):
-        if self.name == 'mpc':
+        if self.name in ['mpc', 'C']:
             return True
         else:
             return False
@@ -97,30 +99,30 @@ class Field(object):
     def tollerance(self):
         if self.name in ['gaussian rational', 'finite field']:
             return 0
-        elif self.name == 'mpc':
+        elif self.name in ['mpc', 'C']:
             return mpmath.mpf('10e-{}'.format(int(min([0.95 * mpmath.mp.dps, mpmath.mp.dps - 4]))))
-        elif self.name == 'padic':
+        elif self.name in ['padic', 'Qp']:
             return PAdic(0, self.characteristic, 0, self.digits)
 
     @property
     def singular_notation(self):
-        if self.name == 'mpc':
+        if self.name in ['mpc', 'C']:
             return '(complex,{},I)'.format(self.digits - 5)
-        elif self.name in ['finite field', 'padic']:
+        elif self.name in ['finite field', 'Fp', 'padic', 'Qp']:
             return str(self.characteristic)
         else:
             return None
 
     def sqrt(self, val):
-        if self.name == "finite field":
+        if self.name in ["finite field", 'Fp']:
             if not isinstance(val, ModP):
                 val = self(val)
             return finite_field_sqrt(val)
-        elif self.name == "padic":
+        elif self.name in ["padic", 'Qp']:
             if not isinstance(val, PAdic):
                 val = self(val)
             return padic_sqrt(val)
-        elif self.name == "mpc":
+        elif self.name in ["mpc", 'C']:
             return mpmath.sqrt(val)
         else:
             raise Exception(f"Field not understood: {self.field.name}")
@@ -131,21 +133,31 @@ class Field(object):
 
     i = I = j
 
-    def random_element(self, shape=(1, )):
+    def random(self, shape=(1, )):
         if shape == (1, ):
-            if self.name == "padic":
+            if self.name in ["padic", 'Qp']:
                 p, k = self.characteristic, self.digits
                 return PAdic(random.randrange(0, p ** k - 1), p, k)
-            elif self.name == "finite field":
+            elif self.name in ["finite field", 'Fp']:
                 p = self.characteristic
                 return ModP(random.randrange(0, p), p)
-            elif self.name == "mpc":
+            elif self.name in ["mpc", 'C']:
                 return mpmath.mpc(str(Q(random.randrange(-100, 101), random.randrange(1, 201))),
                                   str(Q(random.randrange(-100, 101), random.randrange(1, 201))))
             else:
                 raise NotImplementedError
         else:
             raise NotImplementedError
+
+    def random_element(self, *args, **kwargs):
+        import warnings
+        warnings.warn(
+            "random_element is deprecated and will be removed in a future version. "
+            "Use random instead.",
+            DeprecationWarning,
+            stacklevel=2
+        )
+        return self.random(*args, **kwargs)
 
     def epsilon(self, shape=(1, ), ):
         if shape == (1, ):
@@ -165,7 +177,7 @@ class Field(object):
     @property
     def ε(self):
         if not hasattr(self, "_ε"):
-            self._ε = self.epsilon
+            self._ε = self.epsilon()
         return self._ε
 
     @ε.setter